Move ChannelMixin
